+++ src/main/java/me/nobaboy/nobaaddons/mixins/accessors/OverlayTextureAccessor.java
@@ -1,0 +1,12 @@
+package me.nobaboy.nobaaddons.mixins.accessors;
+
+import net.minecraft.client.render.OverlayTexture;
+import net.minecraft.client.texture.NativeImageBackedTexture;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(OverlayTexture.class)
+public interface OverlayTextureAccessor {
+	@Accessor
+	NativeImageBackedTexture getTexture();
+}