+++ src/main/kotlin/me/nobaboy/nobaaddons/events/EventListener.kt
@@ -1,0 +1,8 @@
+package me.nobaboy.nobaaddons.events
+
+/**
+ * Event listener invoked from an [EventDispatcher]
+ */
+fun interface EventListener<T : Event> {
+	fun invoke(event: T)
+}