+++ src/main/kotlin/me/nobaboy/nobaaddons/events/impl/RepoReloadEvent.kt
@@ -3,7 +3,7 @@ package me.nobaboy.nobaaddons.events.impl
 import me.nobaboy.nobaaddons.events.Event
 import me.nobaboy.nobaaddons.events.EventDispatcher
 
-class RepoReloadEvent : Event() {
+class RepoReloadEvent : Event {
 	companion object {
 		val EVENT = EventDispatcher<RepoReloadEvent>()
 	}