+++ src/main/kotlin/me/nobaboy/nobaaddons/events/impl/client/InventoryEvents.kt
@@ -28,13 +28,15 @@ object InventoryEvents {
 	@JvmField val SLOT_UPDATE = EventDispatcher<SlotUpdate>()
 
 	/**
-	 * Event invoked when a slot is clicked in an inventory.
+	 * Event invoked when a slot is clicked while a container is open; this may include slots in the player's inventory.
+	 *
+	 * This may also include clicking out of the GUI, which will typically be a [SlotClick.slot] of `-999`.
 	 */
 	@JvmField val SLOT_CLICK = EventDispatcher<SlotClick>()
 
-	data class Open(val inventory: InventoryData) : Event()
-	data class Close(val sameName: Boolean) : Event()
-	data class Update(val inventory: InventoryData) : Event()
-	data class SlotUpdate(val itemStack: ItemStack, val slot: Int) : Event()
-	data class SlotClick(val itemStack: ItemStack, val button: Int, val slot: Int, val actionType: SlotActionType) : Event()
+	data class Open(val inventory: InventoryData) : Event
+	data class Close(val sameName: Boolean) : Event
+	data class Update(val inventory: InventoryData) : Event
+	data class SlotUpdate(val itemStack: ItemStack, val slot: Int) : Event
+	data class SlotClick(val itemStack: ItemStack, val button: Int, val slot: Int, val actionType: SlotActionType) : Event
 }