+++ src/main/kotlin/me/nobaboy/nobaaddons/features/events/mythological/NearestWarp.kt
@@ -2,5 +2,5 @@ package me.nobaboy.nobaaddons.features.events.mythological
 
 data class NearestWarp(
 	val warpPoint: BurrowWarpLocations.WarpPoint,
-	var used: Boolean = false
+	var used: Boolean = false,
 )