+++ src/main/kotlin/me/nobaboy/nobaaddons/features/inventory/slotinfo/items/VacuumPestsSlotInfo.kt
@@ -11,7 +11,7 @@ import me.nobaboy.nobaaddons.utils.items.ItemUtils.stringLines
 
 // TODO: Add cap option of 40 (which is max fortune), and consider red !!! if >=100k
 object VacuumPestsSlotInfo : ISlotInfo {
-	private val VACUUM_PESTS_REGEX by Regex("Vacuum Bag: (?<amount>\\d+) Pests").fromRepo("slot_info.vacuum_pests")
+	private val VACUUM_PESTS_REGEX by Regex("^Vacuum Bag: (?<amount>[\\d,]+) Pests").fromRepo("slot_info.vacuum_pests")
 
 	private val gardenVacuums = listOf(
 		"SKYMART_VACUUM",