+++ src/main/kotlin/me/nobaboy/nobaaddons/ui/ElementAlignment.kt
@@ -1,7 +1,12 @@
 package me.nobaboy.nobaaddons.ui
 
-enum class ElementAlignment {
+import net.minecraft.util.StringIdentifiable
+
+enum class ElementAlignment : StringIdentifiable {
 	LEFT,
+	CENTER,
 	RIGHT,
 	;
+
+	override fun asString(): String = name
 }